requirements.txt
boto3==1.9.97
Django==2.1.7
django-debug-toolbar==1.11
django-ses==0.8.9
django-storages==1.7.1
djangorestframework==3.9.0
flake8==3.7.6
loggly-python-handler==1.0.0
psycopg2-binary==2.7.7
